+++ CHANGELOG.md
@@ -1,9 +1,25 @@
 # Changelog
 
+## 6.3.7
+
+##### Features
+-  add visible logging for relevant output
+
+##### Fixes
+- ignore ci/cd environment if head SHA hash is not matching environment variable commit hash
+
+
+## 6.3.6
+
+##### Fixes
+- fix value of describe.tag.version placeholder
+
+
 ## 6.3.5
 
 ##### Fixes
 - Improve compatibility with Gradle Configuration Cache #99
+
 
 ## 6.3.4